Moved delete_base and delete_moives to dev_tools app
The dev_tools app will now contain all admin tools.
