+++ dev_tools/static/dev_tools/css/delete_base.css
