+++ dev_tools/static/dev_tools/js/delete_base.js
