+++ dev_tools/static/dev_tools/js/delete_movies.js
