+++ dev_tools/templates/dev_tools/delete_base.html
@@ -1,6 +1,5 @@
 {% load static %}
 {% load static tailwind_tags %}
-
 <!DOCTYPE html>
 <html lang="en">
   <head>
@@ -10,71 +9,38 @@
     <script src="https://kit.fontawesome.com/c60455fa26.js"
             crossorigin="anonymous"></script>
     <link rel="stylesheet" href="{% static 'landing_page/css/base.css' %}" />
-    <link rel="stylesheet" href="{% static 'accounts/css/delete_base.css' %}" />
+    <link rel="stylesheet" href="{% static 'dev_tools/css/delete_base.css' %}" />
     <link rel="stylesheet" href="{% static 'accounts/css/modal.css' %}" />
     {% block head %}
     {% endblock head %}
   </head>
-
   <header>
-    <div id="logo-div">
-      <img src="{% static 'landing_page/images/REELDIN.png' %}"
-           alt="ReeldIn Logo" />
-      <h1>ReeldIn</h1>
-    </div>
-    <div id="profile-div">
-      <a href="{% url 'recommendations:index' %}">
-        {% if user.is_authenticated %}
-          <button id="get-recommendations-btm">Get Recommendations</button>
-        {% else %}
-          <button id="get-recommendations-btm">Demo Recommendations</button>
-        {% endif %}
-      </a>
-      {% if user.is_authenticated %}
-        <a href="{% url 'accounts:logout' %}">
-          <button id="login-button">Sign Out</button>
-        </a>
-      {% else %}
-        <a href="{% url 'accounts:login' %}">
-          <button id="login-button">Sign In</button>
-        </a>
-      {% endif %}
-      {% if user.is_authenticated %}
-        <a href="/profile">
-          <img id="profile-pic" src="{{ request.user.profile_picture.url }}" />
-        </a>
-      {% endif %}
-    </div>
+    {% include 'base/header_base.html' %}
   </header>
-
   <nav>
   </nav>
-
   <body>
     <div class="onboarding-header">
       <h1 id="onboarding-title" class="title-div">
-        {% block header %}
-        {% endblock header %}
+        {% block headerTitle %}
+        {% endblock headerTitle %}
       </h1>
       <p id="onboarding-caption">
         {% block caption %}
         {% endblock caption %}
       </p>
     </div>
-
     <div class="onboarding-content content-div">
       {% block content %}
       {% endblock content %}
     </div>
-
     <!--Forward and backwards buttons-->
     <div class="footer-buttons">
       {% block footer_buttons %}
       {% endblock footer_buttons %}
     </div>
-
     <!-- Exit modal-->
-    <section class="modal hidden" id ="confirm_exit_modal">
+    <section class="modal hidden" id="confirm_exit_modal">
       <div class="flex">
         <h3>
           {% block exit_header %}
@@ -89,9 +55,8 @@
       {% endblock exit_button %}
       <button class="btn" onclick="closeExitModal()">No</button>
     </section>
-    <div class="overlay hidden" id ="exit_overlay" onclick="closeExitModal()"></div>
-
-    <script src="{% static 'accounts/js/delete_base.js' %}"></script>
+    <div class="overlay hidden" id="exit_overlay" onclick="closeExitModal()"></div>
+    <script src="{% static 'dev_tools/js/delete_base.js' %}"></script>
     {% block scripts %}
     {% endblock scripts %}
   </body>
